Bug #33223230 auth_sec.openssl_without_fips fails on big endian platform
The variable underlying option MYSQL_OPT_SSL_FIPS_MODE is
  'ulong opt_ssl_fips_mode'
so access it as such, rather than uint.

MEB uses the same option, with an enum to hold the value. Convert it
to ulong before trying to set FIPS mode.

(cherry picked from commit 3deb3bb3cc588dc48162f40cbb1d6176ee07c4c5)

+++ sql-common/client.cc
@@ -8235,7 +8235,8 @@ int STDCALL mysql_options(MYSQL *mysql, enum mysql_option option,
     case MYSQL_OPT_SSL_FIPS_MODE: {
       char ssl_err_string[OPENSSL_ERROR_LENGTH] = {'\0'};
       ENSURE_EXTENSIONS_PRESENT(&mysql->options);
-      mysql->options.extension->ssl_fips_mode = *static_cast<const uint *>(arg);
+      mysql->options.extension->ssl_fips_mode =
+          *static_cast<const ulong *>(arg);
       if (set_fips_mode(mysql->options.extension->ssl_fips_mode,
                         ssl_err_string) != 1) {
         DBUG_PRINT("error", ("fips mode set error %s:", ssl_err_string));
